fix(ui): static mac sidebar branding and Top Genre deeplink
